LPS-77946 Move deleteOrphanedAssetEntries from verify process to 7.0 upgrade
