+++ modules/apps/web-experience/asset/asset-service/src/main/java/com/liferay/asset/internal/verify/AssetServiceVerifyProcess.java
@@ -14,20 +14,11 @@
 
 package com.liferay.asset.internal.verify;
 
-import com.liferay.asset.kernel.service.AssetEntryLocalService;
-import com.liferay.document.library.kernel.model.DLFileEntryConstants;
-import com.liferay.document.library.kernel.service.DLFileEntryLocalService;
-import com.liferay.portal.kernel.dao.orm.EntityCacheUtil;
-import com.liferay.portal.kernel.dao.orm.FinderCacheUtil;
 import com.liferay.portal.kernel.util.LoggingTimer;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.verify.VerifyLayout;
 import com.liferay.portal.verify.VerifyProcess;
-import com.liferay.portlet.asset.model.impl.AssetEntryImpl;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Douglas Wong
@@ -39,29 +30,8 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class AssetServiceVerifyProcess extends VerifyLayout {
 
-	protected void deleteOrphanedAssetEntries() throws Exception {
-		try (LoggingTimer loggingTimer = new LoggingTimer()) {
-			long classNameId = _portal.getClassNameId(
-				DLFileEntryConstants.getClassName());
-
-			StringBundler sb = new StringBundler(5);
-
-			sb.append("delete from AssetEntry where classNameId = ");
-			sb.append(classNameId);
-			sb.append(" and classPK not in (select fileVersionId from ");
-			sb.append("DLFileVersion) and classPK not in (select fileEntryId ");
-			sb.append("from DLFileEntry)");
-
-			runSQL(sb.toString());
-
-			EntityCacheUtil.clearCache(AssetEntryImpl.class);
-			FinderCacheUtil.clearCache(AssetEntryImpl.class.getName());
-		}
-	}
-
 	@Override
 	protected void doVerify() throws Exception {
-		deleteOrphanedAssetEntries();
 		verifyAssetLayouts();
 	}
 
@@ -70,14 +40,5 @@ public class AssetServiceVerifyProcess extends VerifyLayout {
 			verifyUuid("AssetEntry");
 		}
 	}
-
-	@Reference
-	private AssetEntryLocalService _assetEntryLocalService;
-
-	@Reference
-	private DLFileEntryLocalService _dlFileEntryLocalService;
-
-	@Reference
-	private Portal _portal;
 
 }
